requirements.txt
streamlit
langchain
langchain-community
langchain-groq
groq
python-dotenv
